Added new features
Added copyright notes to all the files, and to any jar file from now on.

Added an option to make creeper explosion insta break player shield if he block.

Added a damage bypass for creeper explosion, in %, ranging from 0 to 1, in which the damage from a creeper explosion will bypass the player shield protection.

And also added an option to send player a message via chat/title when that happens.